# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This API is for tracking your finance",
        "title": "Finance Tracker API",
        "contact": {
            "name": "API Support",
            "email": "teguh180902@gmail.com"
        },
        "license": {
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/register": {
            "post": {
                "description": "Usage of this endpoint will Create User For Login",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Register User",
                "parameters": [
                    {
                        "description": "User login information",
                        "name": "loginRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login success",
                        "schema": {
                            "$ref": "#/definitions/user.SuccessRegister"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/user.FailureRegisterResponse"
                        }
                    }
                }
            }
        },
...
        "/user": {
            "get": {
                "description": "Usage of this endpoint is for gett all user without jwt param",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get All data",
                "responses": {
                    "200": {
                        "description": "Succes Get All Data",
                        "schema": {
                            "$ref": "#/definitions/user.GetAllUser"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/user.FailureResponse"
                        }
                    }
                }
            }
        },
        "/user/{id}": {
            "get": {
                "description": "Usage of this endpoint is for gett all user without jwt param",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Update User",
                "parameters": [
                    {
                        "description": "User login information",
                        "name": "UpdateRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UserUpdateRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Jwt Token",
                        "name": "x-token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "id user to update",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succes Update Data",
                        "schema": {
                            "$ref": "#/definitions/user.GetAllUser"
                        }
                    },
                    "400": {
                        "description": "Update data failed",
                        "schema": {
                            "$ref": "#/definitions/user.FailureResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "entity.Acount": {
            "type": "object",
            "properties": {
                "acount_name": {
                    "type": "string"
                },
                "balance": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "deletedat": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "transaction": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.Transaction"
                    }
                },
                "updatedat": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "entity.Transaction": {
            "type": "object",
            "properties": {
                "acount_id": {
                    "type": "integer"
                },
                "amount": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "deletedat": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "transaction_type": {
                    "type": "string"
                },
                "updatedat": {
                    "type": "string"
                }
            }
        },
        "entity.User": {
            "type": "object",
            "properties": {
                "acounts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.Acount"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "deletedat": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "updatedat": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "request.User": {
            "type": "object",
            "required": [
                "email",
                "name",
                "password",
                "username"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "teguh22@gmail.com"
                },
                "name": {
                    "type": "string",
                    "example": "guhh"
                },
                "password": {
                    "type": "string",
                    "example": "password123"
                },
                "username": {
                    "type": "string",
                    "example": "guhh"
                }
            }
        },
        "request.UserUpdateRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "user.FailureRegisterResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Email or Password is Already Exist"
                },
                "succes": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "user.FailureResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Can't get data"
                },
                "succes": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "user.GetAllUser": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.User"
                    }
                },
                "message": {
                    "type": "string"
                },
                "succes": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "user.SuccessRegister": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/entity.User"
                },
                "message": {
                    "type": "string"
                },
                "succes": {
                    "type": "boolean",
                    "example": true
                }
            }
        }
    }
}